Fix revalidation oom (#313)
* Fix spammer timestamp

* Improve strict semantics check

* Abort revalidation walk on milestones to prevent OOM

When revalidation kicked in and walked a cone, it happened that the cones
were so big that the node crashed with OOM. This is fixed by aborting the
walk if a milestone is found, and restarting the walk at the found milestone.

* Fix revalidation/snapshot import order

If a snapshot was not imported successfully, the revalidation was started at next
node startup and crashed. This was misleading and is fixed.