Release v0.5.3 rc1 (#650)
* Fix missing trytes conversion

* Dashboard improvements

This introduces a new SyncStatus message to reduce the websocket traffic during syncing. It also removes all workerpools because the websocket hub already has a bufferend channel for all messages. Value transactions are now filtered before they get rate limited.

* Update hive.go to prevent shutdown deadlock

* Do not exit pruning thread if history not enough

* Change pruning logic to delete all tx that were referenced by a milestone

* Update go modules

* Fix typos

* Add details to pruning error message

* Add testsuite to make the creation of node testcases easier

* Feed the dog

* Close connection on connection error

* Flush storages in every revalidation step.

This is important, otherwise the txHashes could still exist in the database in the moment we check for existence, which leads the pruning logic to fail.

* Fix cleanupApprovers in revalidation

* Update go modules

* Add different flagSets for different configs.

This was necessary to fix a bug where the global default settings were written to the peering.json because all CommandLine flags were bind to it.

* Show correct alias for incoming connections

* Ignore autopeers in CfgPeeringMaxPeers check

* Fix parsing of flags

* Release v0.5.3-rc1