Fix INX deadlock (#1560)
* Add ReferencedBlocksCountUpdated event

* Remove unused tangle events

* Pass options to INX workerpools

* Decouple firing of several tangle events from ledger write lock.

It is safe to fire several confirmation related events in the tangle package without holding the ledger lock. Comments were added to the events for better understanding. This was done in the first place to fix a race condition in INX, where the write lock of the ledger was deadlocked with the internal lock of gRPC do to asynchronous calls to outputById which acquires the ledger read lock.